I created a method, parametersToFile, in Controller to write the parameters used to initialize a given run to an output file. Also in Controller I created methods: writeTeamToFile and writeHeadersToFile simply to avoid further duplicating code in the Controller for writing teams to file.
In main I added OUTPUT_FILE_PATH so that output can be directed to a different folder than input. This is now passed to the Controller as an argument. I also added a variable to simulation specifying the name of an output file.
In main I added a call to controller's parametersToFile to write the parameters used to initialize the simulation to file.